refactor(ws): extract device-cap eviction into tabs_persist_retention module
Pure move: tabs_persist.rs sits at the 1,000-line repo limit and the
upcoming defect fixes need room. No behavior change.
